+++ db/migrate/20150710164631_add_stratifies_by_site_to_stratification_factors.rb
@@ -1,0 +1,5 @@
+class AddStratifiesBySiteToStratificationFactors < ActiveRecord::Migration
+  def change
+    add_column :stratification_factors, :stratifies_by_site, :boolean, null: false, default: false
+  end
+end