+++ db/migrate/20150713144442_add_extra_options_to_lists.rb
@@ -1,0 +1,5 @@
+class AddExtraOptionsToLists < ActiveRecord::Migration
+  def change
+    add_column :lists, :extra_options, :text
+  end
+end